backend/pyproject.toml
[project]
name = "emotion-song-recommendation-backend"
version = "1.1.0"
requires-python = ">=3.9,<3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "tensorflow==2.10.*",
    "keras",
    "opencv-python",
    "numpy<2",
    "requests",
    "python-multipart",
    "python-dotenv",
]
